requirements.txt
astropy==1.3.3
beautifulsoup4==4.5.3
bs4==0.0.1
cmocean==1.1
cycler==0.10.0
DateTime==4.1.1
matplotlib==1.5.3
numpy==1.11.3
pandas==0.19.2
patsy==0.4.1
pykalman==0.9.5
pyparsing==2.1.9
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
requests==2.13.0
scipy==0.18.1
six==1.10.0
statsmodels==0.6.1
zope.interface==4.3.2
